Fixed both graphing bugs.
Both reflection about the x axis and only graphing in yMin to yMax range
in pixels.

The reflection issue was fixed by subtracting the data points y Value
from the max yValue in pixels.  To fix the only graphing from yMin to
yMax in pixel i computed a scaling  factor and multiplied the computed y
value by it